implement creating of an entry
